+++ iop/hdd/apa/src/hdd_fio.h
@@ -14,7 +14,7 @@
 // I/O functions
 int hddInit(iop_device_t *f);
 int hddDeinit(iop_device_t *f);
-int hddFormat(iop_file_t *f, const char *dev, const char *blockdev, void *arg, size_t arglen);
+int hddFormat(iop_file_t *f, const char *dev, const char *blockdev, void *arg, int arglen);
 int hddOpen(iop_file_t *f, const char *name, int flags, int mode);
 int hddClose(iop_file_t *f);
 int hddRead(iop_file_t *f, void *buf, int size);
