+++ iop/system/iomanx/src/iomanX.c
@@ -258,7 +258,7 @@ int close(int fd)
 	return res;
 }
 
-int read(int fd, void *ptr, size_t size)
+int read(int fd, void *ptr, int size)
 {
 	iop_file_t *f = get_file(fd);
 
@@ -268,7 +268,7 @@ int read(int fd, void *ptr, size_t size)
 	return f->device->ops->read(f, ptr, size);
 }
 
-int write(int fd, void *ptr, size_t size)
+int write(int fd, void *ptr, int size)
 {
 	iop_file_t *f = get_file(fd);
 
@@ -291,7 +291,7 @@ int lseek(int fd, int offset, int whence)
 	return f->device->ops->lseek(f, offset, whence);
 }
 
-int ioctl(int fd, unsigned long cmd, void *arg)
+int ioctl(int fd, int cmd, void *arg)
 {
 	iop_file_t *f = get_file(fd);
 
@@ -496,7 +496,7 @@ int chstat(const char *name, iox_stat_t *stat, unsigned int mask)
     return file.device->ops->chstat(&file, filename, stat, mask);
 }
 
-int format(const char *dev, const char *blockdev, void *arg, size_t arglen)
+int format(const char *dev, const char *blockdev, void *arg, int arglen)
 {
 	iop_file_t file;
 	char *filename;
@@ -550,7 +550,7 @@ int sync(const char *dev, int flag)
 	return path_common(dev, flag, 0x106);
 }
 
-int mount(const char *fsname, const char *devname, int flag, void *arg, size_t arglen)
+int mount(const char *fsname, const char *devname, int flag, void *arg, int arglen)
 {
 	iop_file_t file;
 	char *filename;
@@ -596,7 +596,7 @@ long long lseek64(int fd, long long offset, int whence)
 	return f->device->ops->lseek64(f, offset, whence);
 }
 
-int devctl(const char *name, int cmd, void *arg, size_t arglen, void *buf, size_t buflen)
+int devctl(const char *name, int cmd, void *arg, unsigned int arglen, void *buf, unsigned int buflen)
 {
 	iop_file_t file;
 	char *filename;
@@ -616,7 +616,7 @@ int symlink(const char *old, const char *new)
 	return link_common(old, new, 8);
 }
 
-int readlink(const char *name, char *buf, size_t buflen)
+int readlink(const char *name, char *buf, unsigned int buflen)
 {
 	iop_file_t file;
 	char *filename;
@@ -632,7 +632,7 @@ int readlink(const char *name, char *buf, size_t buflen)
 }
 
 
-int ioctl2(int fd, int command, void *arg, size_t arglen, void *buf, size_t buflen)
+int ioctl2(int fd, int command, void *arg, unsigned int arglen, void *buf, unsigned int buflen)
 {
 	iop_file_t *f;
